--- client/index.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>UMass Amherst CICS Courses Visualizer</title>
-  <script src="https://d3js.org/d3.v7.min.js"></script>
-  <script type="module" src="js/simulation.js" defer></script>
-  <!-- Bootstrap inclusion via CDN -->
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet"
-    integrity="sha384-9ndCyUaIbzAi2FUVXJi0CjmCapSmO7SnpJef0486qhLnuZ2cdeRhO02iuK6FUUVM" crossorigin="anonymous">
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-geWF76RCwLtnZ8qwWowPQNguL3RmwHVBC9FhGdlKrxdiJJigb/j/68SIy3Te4Bkz" crossorigin="anonymous"
-    defer></script>
-  <!-- Google Fonts -->
-  <link rel="preconnect" href="https://fonts.googleapis.com">
-  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-  <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:wght@400&display=swap" rel="stylesheet">
-  <!-- Personal stylesheet  -->
-  <link rel="stylesheet" href="css/styles.css">
-</head>
-
-<body>
-  <nav class="navbar bg-dark border-bottom border-bottom-dark" data-bs-theme="dark"">
-        <div class=" container-fluid">
-    <a class="navbar-brand">UMass CICS Course Visualizer</a>
-    <form class="d-flex" role="search">
-      <input class="form-control me-2" type="search" placeholder="Search" aria-label="Search">
-      <button class="btn btn-outline-success" type="submit">Search</button>
-    </form>
-    </div>
-  </nav>
-
-  <svg id="svg" viewBox="0 0 2000 2000" width="2000" height="2000">
-    <g id="container"></g>
-  </svg>
-</body>
-
-</html>
