+++ umass-cics-course-visualizer/src/App.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import CourseVisualizer from './CourseVisualizer';
+
+function App() {
+  return (
+    <div className="App">
+      <CourseVisualizer />
+    </div>
+  );
+}
+
+export default App;
