+++ umass-cics-course-visualizer/src/CourseVisualizer.css
@@ -1,5 +1,4 @@
-html,
-body {
+.course-visualizer {
     margin: 0;
     height: 100%;
     overflow: hidden;
@@ -7,7 +6,7 @@ body {
     background-color: #f5f5f7;
 }
 
-#svg{
+svg {
     height: 100vh;
     width: 100vw;
     display: block;
