+++ umass-cics-course-visualizer/src/util/Util.js
@@ -7,32 +7,16 @@ async function fetchCourseID(search) {
             const response = await fetch(`https://spire-api.melanson.dev/courses/?page=${page}&search=${search}`);
             if (!response.ok) throw new Error(`HTTP error: ${response.status}`);
             const data = await response.json();
-            //
-            // Combine the new data with the results we have so far.
-            // let updated = data.filter(x => {
-            //     x["offerings"].forEach(e => {
-            //        if(e.hasOwnProperty("term") && e["term"].hasOwnProperty("id") && e["term"]["id"] === "Spring 2023"){
-            //               return true; 
-            //        }
-            //     });
-            // });
             results = results.concat(data);
 
-            // Check if there's a next page. This will depend on how your API communicates this.
-            // If the API includes a 'next' field in the response that is null when there are no more pages:
             if (!data.next) break;
-
-            // If the API includes a 'total_pages' field in the response:
-            // if (page >= data.total_pages) break;
-
-            // Go to the next page
             page++;
         } catch (error) {
             console.log("could not get data");
             break;
         }
     }
-    //console.log(results)
+
     return results;
 }
 
@@ -62,7 +46,7 @@ function findCourses(text) {
     return result;
 }
 
+export { fetchCourseID, findCourses };
 
-export{ fetchCourseID, findCourses };
